Use prev_dp and drop the first dimension of the dp
The recurrence relation for calculating dp[m][...] never
needs to reach back further than dp[m-1][...], so it suffices to
keep just the dp[] values for the previous m value.
We can achieve this by dropping the first dimension of the dp,
and using prev_dp for storing what we used to store in dp[m].
Doing this drops space usage from O(n) to O(1).

https://stackoverflow.com/questions/67240898/codejam-2021-qualifier-round-moons-and-umbrellas-algorithm-explanation